+++ crates/vm/levm/src/errors.rs
@@ -196,6 +196,9 @@ pub enum InternalError {
     AccountNotDelegated,
     #[error("No recipient found for privilege transaction")]
     RecipientNotFoundForPrivilegeTransaction,
+    //TODO: Refactor all errors. https://github.com/lambdaclass/ethrex/issues/2886
+    #[error("Custom error: {0}")]
+    Custom(String),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, thiserror::Error, Serialize, Deserialize)]